10:12 AM updated LSCVDist to use pointer to ColPivHouseholderQR<MatrixX2d>; 12:09 PM fixed the remaining pointers in LSCVDist and renamed getDistPtr to getDistFunc; 12:29 PM removed the address pointers from MI and CCRE too and also removed the underscore from the erstwhile only private members of these classes; 1:39 PM made all remaining dist member variables const; added getSPIMask to SearchMethod and made spi_mask protected; 2:42 PM got several new tors; 3:35 PM completed initial version of LieAffine; 3:43 PM added it to mtf.h and its parameters to params; 3:45 PM added it to register.h; 3:47 PM added it to the make and cmake build systems;

+++ Examples/src/generateSyntheticSeq.cc
@@ -109,8 +109,8 @@ int main(int argc, char * argv[]) {
 
 	double font_size = 1.00;
 
-	resx = size_x;
-	resy = size_y;
+	resx = static_cast<unsigned int>(size_x);
+	resy = static_cast<unsigned int>(size_y);
 
 	mtf::SSM ssm(mtf::getSSM(syn_ssm.c_str()));
 	if(!ssm){
